refactor: move around UI to reduce massive file
