+++ packages/snap/src/transaction-ui/transaction-title.ts
@@ -1,0 +1,89 @@
+import { invalidParameters } from '../errors';
+
+/**
+ * Formats a human readable transaction title based on the transaction command.
+ *
+ * @param tx - Object with a single command property. Uusally the incoming `transaction` property from `client.send_transaction`.
+ * @returns A human readable transaction title.
+ */
+export function transactionTitle(tx: any): string {
+  const keys = Object.keys(tx);
+
+  if (keys.length !== 1) {
+    throw invalidParameters('Invalid transaction');
+  }
+
+  switch (keys[0]) {
+    case 'orderSubmission':
+      return 'Order submission';
+    case 'orderCancellation':
+      return 'Order cancellation';
+    case 'orderAmendment':
+      return 'Order amendment';
+    case 'withdrawSubmission':
+      return 'Withdraw submission';
+    case 'proposalSubmission':
+      return 'Proposal submission';
+    case 'voteSubmission':
+      return 'Vote submission';
+    case 'liquidityProvisionSubmission':
+      return 'Liquidity provision';
+    case 'delegateSubmission':
+      return 'Delegate submission';
+    case 'undelegateSubmission':
+      return 'Undelegate submission';
+    case 'liquidityProvisionCancellation':
+      return 'Liquidity provision cancellation';
+    case 'liquidityProvisionAmendment':
+      return 'Liquidity provision amendment';
+    case 'transfer':
+      return 'Transfer';
+    case 'cancelTransfer':
+      return 'Cancel transfer';
+    case 'announceNode':
+      return 'Announce node';
+    case 'batchMarketInstructions':
+      return 'Batch market instructions';
+    case 'stopOrdersSubmission':
+      return 'Stop orders submission';
+    case 'stopOrdersCancellation':
+      return 'Stop orders cancellation';
+    case 'nodeVote':
+      return 'Node vote';
+    case 'nodeSignature':
+      return 'Node signature';
+    case 'chainEvent':
+      return 'Chain event';
+    case 'keyRotateSubmission':
+      return 'Key rotation submission';
+    case 'stateVariableProposal':
+      return 'State variable proposal';
+    case 'validatorHeartbeat':
+      return 'Validator heartbeat';
+    case 'ethereumKeyRotateSubmission':
+      return 'Ethereum key rotation submission';
+    case 'protocolUpgradeProposal':
+      return 'Protocol upgrade proposal';
+    case 'issueSignatures':
+      return 'Issue signatures';
+    case 'oracleDataSubmission':
+      return 'Oracle data submission';
+    case 'createReferralSet':
+      return 'Create referral set';
+    case 'updateReferralSet':
+      return 'Update referral set';
+    case 'applyReferralCode':
+      return 'Apply referral code';
+    case 'batchProposalSubmission':
+      return 'Batch proposal submission';
+    case 'updatePartyProfile':
+      return 'Update party profile';
+    case 'updateMarginMode':
+      return 'Update margin mode';
+    case 'joinTeam':
+      return 'Join team';
+    default:
+      // TODO: Should this throw? If the protos are out of date this will prevent new transactions being sent. Should this be a warning?
+      throw invalidParameters('Unknown transaction type');
+  }
+}
